# docs/09-api/openapi.yaml
openapi: 3.0.3
info:
  title: Wovyr AI Platform API
  version: "0.3.0"
  description: |
    The Wovyr single-node server's HTTP API, as actually implemented by
    `crates/wovyr-server` (not the aspirational conventions in `overview.md` —
    see the note below). Source of truth: this spec is generated by hand from
    the Axum route handlers, kept in sync manually until a codegen pipeline
    exists.

    ## Divergence from `overview.md`

    `overview.md` describes a target-state convention (opaque `agt_01H...`
    ids, OAuth2/JWT/mTLS auth, a generic `/operations/{id}` polling resource,
    `RateLimit-*` headers, a uniform `{id, object, tenant, project,
    created_at, updated_at, version}` envelope on every resource). **None of
    that is implemented today.** The real API:

    - Identifies resources by their natural key (agent name, workflow
      `execution_id`, secret `name`, `publisher/name` for plugins/listings) —
      not a generated opaque id.
    - Authenticates via a verified `Authorization: Bearer <credential>`
      (RM-GA-P1 SEC-101, `crates/wovyr-server/src/auth.rs`): `WOVYR_AUTH_MODE=jwt`
      checks a JWT's signature (HS256/RS256) plus expiry/issuer/audience, or
      `apikey` looks up a SHA-256-hashed bearer token — either way the
      verified principal overwrites `X-Wovyr-Principal` before RBAC runs, so a
      client-asserted header alone can no longer win. No OAuth2 authorization
...
      are minted by an external IdP; API keys are minted via
...
servers:
...
components:
  securitySchemes:
    tenantHeader:
      type: apiKey
      in: header
      name: X-Wovyr-Tenant
      description: |
        The acting tenant. Defaults to `"default"` (anonymous/back-compat,
        unrestricted) if absent. Not cryptographically verified — this header
        is a claim, authorized against real memberships server-side.
    principalHeader:
      type: apiKey
      in: header
      name: X-Wovyr-Principal
      description: |
        The acting principal id, used verbatim only in the default
        `disabled-loopback` auth mode. When `bearerAuth` is configured
        (`WOVYR_AUTH_MODE=jwt|apikey`), the server overwrites this with the
        verified principal before any handler runs — a client-supplied value
        is then ignored, not merged.
    bearerAuth:
      type: http
      scheme: bearer
      description: |
        RM-GA-P1 SEC-101. Present only when the server is configured with
        `WOVYR_AUTH_MODE=jwt` (a signed JWT — HS256 via `WOVYR_JWT_HS_SECRET` or
        RS256 via `WOVYR_JWT_RS_PUBLIC_KEY`, `sub` claim becomes the principal)
        or `WOVYR_AUTH_MODE=apikey` (a bearer token minted via
        `wovyr auth create-key`, SHA-256-hashed server-side). The default
        `disabled-loopback` mode carries no bearer credential at all and
        instead trusts `X-Wovyr-Principal`/`X-Wovyr-Tenant` verbatim — a
        local-dev fallback the server refuses to expose on a non-loopback
        bind unless `WOVYR_ALLOW_ANONYMOUS=1` is set.
  parameters:
...
    xWovyrTenant:
      name: X-Wovyr-Tenant
      in: header
      schema: { type: string, default: default }
...
    xWovyrProject:
      name: X-Wovyr-Project
      in: header
      schema: { type: string }
      description: In-scope project for run-path quota metering.
    idempotencyKey:
      name: Idempotency-Key
      in: header
      schema: { type: string }
      description: >-
        RM-GA-P4 API-703: a repeated key on this route replays the original
        (tenant, method, path, key)-scoped response instead of re-executing
        the mutation. Only a successful response is cached.
...
  schemas:
...
    RunRequest:
      type: object
      required: [manifest]
      properties:
        manifest: { type: string, description: Inline agent YAML manifest. }
        input: { description: Arbitrary JSON input passed to the agent. }
        max_steps: { type: integer, nullable: true }
    RunResult:
      type: object
      properties:
        run_id: { type: string }
        status: { type: string, enum: [succeeded] }
        output: { type: object, properties: { message: { type: string } } }
        steps: { type: integer }
        usage: { $ref: '#/components/schemas/Usage' }
...
paths:
...
  /api/v1/agents:run:
    post:
      tags: [agents]
      summary: Run an inline agent manifest (no persistence)
      parameters:
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrProject'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/RunRequest' } }
      responses:
        '200':
          description: Run result
          content:
            application/json: { schema: { $ref: '#/components/schemas/RunResult' } }
        '400': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }
  /api/v1/agents:stream:
    post:
      tags: [agents]
      summary: Run an inline agent manifest, streaming progress via SSE
      parameters:
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrProject'
      requestBody:
        required: true
        content:
          application/json: { schema: { $ref: '#/components/schemas/RunRequest' } }
      responses:
        '200':
          description: |
            `text/event-stream`. Unnamed `data:` frames in order:
            `{"type":"start",...}`, optional `{"type":"memory",...}`,
            repeated `{"type":"delta","text":...}`, optional
            `{"type":"reasoning","text":...}` (the model's thinking channel,
            where the provider exposes one) and
            `{"type":"tool_call_delta","index":...,"name":...,"arguments":...}`
            (incremental argument fragments as the model composes a call),
            optional `{"type":"tool_call",...}`/`{"type":"tool_result",...}`,
            `{"type":"done","usage":{...}}`. Terminal **named** event
            `event: result` (`data`: run result) or `event: error` (plain-text
            data). Consumers should ignore frame types they don't recognize.
          content:
            text/event-stream: { schema: { type: string } }
        '429': { $ref: '#/components/responses/Error' }
  /api/v1/agents:
    get:
      tags: [agents]
      summary: List stored agents (ids only) for the caller's tenant
      parameters:
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
      responses:
        '200':
          description: Paginated agent ids
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    properties: { data: { type: array, items: { type: string } } }
    post:
      tags: [agents]
      summary: Store an agent manifest
      parameters:
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifest]
              properties: { manifest: { type: string } }
      responses:
        '200':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties: { id: { type: string }, status: { type: string, enum: [created] } }
        '400': { $ref: '#/components/responses/Error' }
  /api/v1/agents/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema: { type: string }
      - $ref: '#/components/parameters/xWovyrTenant'
      - $ref: '#/components/parameters/xWovyrPrincipal'
    get:
      tags: [agents]
      summary: Fetch a stored agent's manifest
      responses:
        '200':
          description: Agent
          content:
            application/json:
              schema:
                type: object
                properties: { id: { type: string }, manifest: { type: string } }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      tags: [agents]
      summary: Delete a stored agent
      parameters:
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '204': { description: Deleted }
        '404': { $ref: '#/components/responses/Error' }
  /api/v1/agents/{id}/run:
    post:
      tags: [agents]
      summary: Run a stored agent by id
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
        - $ref: '#/components/parameters/xWovyrProject'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: { input: {}, max_steps: { type: integer, nullable: true } }
      responses:
        '200':
          description: Run result
          content:
            application/json: { schema: { $ref: '#/components/schemas/RunResult' } }
        '404': { $ref: '#/components/responses/Error' }
        '429': { $ref: '#/components/responses/Error' }

  # ------------------------------------------------------------ workflows --
...
  /api/v1/workflows:
    get:
      tags: [workflows]
      summary: List workflow executions visible to the caller's tenant
      parameters:
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
        - $ref: '#/components/parameters/limit'
        - $ref: '#/components/parameters/cursor'
        - name: workflow
          in: query
          schema: { type: string }
          description: Filter by workflow name.
        - name: status
          in: query
          schema:
            type: string
            enum: [created, validated, scheduled, running, waiting, resumed, compensating, completed, failed, cancelled]
      responses:
        '200':
          description: Paginated execution summaries
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/PageMeta'
                  - type: object
                    properties: { data: { type: array, items: { type: object } } }
        '400': { $ref: '#/components/responses/Error' }
    post:
      tags: [workflows]
      summary: Submit (durably start + async-drive) a workflow execution
      parameters:
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
        - $ref: '#/components/parameters/idempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [manifest]
              properties:
                manifest: { type: string }
                input: {}
                execution_id: { type: string, nullable: true, description: 'Auto-generated `<workflow-name>-<n>` if omitted.' }
      responses:
        '200':
          description: Submitted (does not wait for completion — poll `GET /api/v1/workflows/{id}`)
          content:
            application/json:
              schema:
                type: object
                properties: { execution_id: { type: string }, status: { type: string, enum: [submitted] } }
  /api/v1/workflows/{id}:
    get:
      tags: [workflows]
      summary: Get an execution's live status + event history
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
      responses:
        '200':
          description: Execution + events
          content:
            application/json:
              schema:
                type: object
                properties: { execution: { type: object }, events: { type: array, items: { type: object } } }
        '404': { $ref: '#/components/responses/Error' }
    delete:
      tags: [workflows]
      summary: Cancel an execution (advisory — in-flight activities still complete)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - $ref: '#/components/parameters/xWovyrTenant'
        - $ref: '#/components/parameters/xWovyrPrincipal'
        - $ref: '#/components/parameters/idempotencyKey'
      responses:
        '202': { description: Cancellation accepted }
        '404': { $ref: '#/components/responses/Error' }
...
security:
  # Alternatives (a client uses whichever the server is actually configured
  # for — `WOVYR_AUTH_MODE`): the local-dev default (raw, unverified headers),
  # or a verified bearer credential (RM-GA-P1 SEC-101) alongside the tenant
  # header (still asserted, but authorized against real memberships).
  - tenantHeader: []
    principalHeader: []
  - tenantHeader: []
    bearerAuth: []